Add key to show main terminal (incl accrued log messages)
Tried a bunch of things to do this before realizing my approaches were
overkill. If, instead of wanting the user to be able to press any key to
return, I just print some "press return" text that gets cleaned up on
exit, I can just use stdin.read() instead of getch().

+++ src/log.zig
@@ -57,7 +57,7 @@ pub fn openLogfile() !void {
     logfile = try std.fs.cwd().createFile(logfile_path, .{ .read = true });
 }
 
-pub fn closeLogfile() void {
+pub fn printLogfile() void {
     if (logfile) |file| {
         const stderr = std.io.getStdErr().writer();
         file.seekTo(0) catch {};
@@ -69,6 +69,13 @@ pub fn closeLogfile() void {
             stderr.writeAll(buffer[0..len]) catch {};
             if (len < buffer_size) break;
         }
+        file.setEndPos(0) catch {};
+    }
+}
+
+pub fn closeLogfile() void {
+    printLogfile();
+    if (logfile) |file| {
         file.close();
         std.fs.cwd().deleteFile(logfile_path) catch {};
     }
